Solve circular import
